requirement.txt
# Core Libraries
torch>=2.0.0
transformers>=4.30.0
langchain>=0.0.200
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3  # Use faiss-gpu if CUDA is available
scikit-learn>=1.2.0
pandas>=1.5.0
numpy>=1.23.0

# Data Pipeline
scrapy>=2.8.0
beautifulsoup4>=4.12.0
tweepy>=4.12.0
newsapi-python>=0.2.7
kafka-python>=2.0.2
pymongo>=4.3.3
python-dotenv>=1.0.0

# Backend & Integrations
fastapi>=0.95.0
uvicorn>=0.21.0
twilio>=8.3.0
requests>=2.28.0
redis>=4.5.0

# Sarcasm/Sentiment Models (optional)
aiohttp>=3.8.0
elasticsearch>=8.7.0  # For RAG alternatives

# Testing & Monitoring
locust>=2.15.0
selenium>=4.8.0
prometheus-client>=0.16.0